@@ -12,20 +12,21 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "4.0.0-beta.7",
-    "@angular/compiler": "4.0.0-beta.7",
-    "@angular/core": "4.0.0-beta.7",
-    "@angular/flex-layout": "^2.0.0-beta.5",
-    "@angular/forms": "4.0.0-beta.7",
-    "@angular/http": "4.0.0-beta.7",
-    "@angular/material": "^2.0.0-beta.1",
-    "@angular/platform-browser": "4.0.0-beta.7",
-    "@angular/platform-browser-dynamic": "4.0.0-beta.7",
-    "@angular/router": "4.0.0-beta.7",
-    "@ngrx/core": "^1.2.0",
-    "@ngrx/effects": "^2.0.0",
-    "@ngrx/store": "^2.2.1",
-    "@ngrx/store-devtools": "^3.2.3",
+    "@angular/animations": "^5.2.10",
+    "@angular/cdk": "^5.2.5",
+    "@angular/common": "5.2.10",
+    "@angular/compiler": "5.2.10",
+    "@angular/core": "5.2.10",
+    "@angular/flex-layout": "^5.0.0-beta.14",
+    "@angular/forms": "5.2.10",
+    "@angular/http": "5.2.10",
+    "@angular/material": "^5.2.5",
+    "@angular/platform-browser": "5.2.10",
+    "@angular/platform-browser-dynamic": "5.2.10",
+    "@angular/router": "5.2.10",
+    "@ngrx/effects": "^5.2.0",
+    "@ngrx/store": "^5.2.0",
+    "@ngrx/store-devtools": "^5.2.0",
     "amazon-cognito-identity-js": "^1.13.0",
     "core-js": "^2.4.1",
     "crypto-js": "^3.1.9-1",
@@ -34,8 +35,8 @@
     "zone.js": "^0.7.2"
   },
   "devDependencies": {
-    "@angular/cli": "1.0.0-beta.31",
-    "@angular/compiler-cli": "4.0.0-beta.7",
+    "@angular/cli": "^1.7.4",
+    "@angular/compiler-cli": "5.2.10",
     "@types/crypto-js": "^3.1.33",
     "@types/jasmine": "2.5.38",
     "@types/node": "^6.0.42",
@@ -51,7 +52,7 @@
     "protractor": "~5.1.0",
     "ts-node": "1.2.1",
     "tslint": "^4.3.0",
-    "typescript": "^2.1.6",
+    "typescript": ">=2.4.2 <2.7.0",
     "wallaby-webpack": "^0.0.33"
   }
 }
