@@ -1,9 +1,8 @@
-import { Component, OnInit, OnDestroy, Input, Output, EventEmitter } from '@angular/core';
+import { Component, OnInit, OnDestroy } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
 import { Subject } from 'rxjs/Subject';
 
 import { ValidationT, ValidT, MobileNumberT } from '../../services/form-validation/form-validation.service';
-import { Animations } from '../animations';
 
 import { Store } from '@ngrx/store';
 import { AppStateT } from '../../../../../ngrx';
@@ -15,8 +14,6 @@ import * as fromCredential from '../../../../../ngrx/auth/credential/credential.
   selector: 'msp-forgot-password-widget',
   templateUrl: './forgot-password-widget.component.html',
   styleUrls: ['./forgot-password-widget.component.css'],
-  animations: [Animations.swipeInOut],
-  // changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class ForgotPasswordWidgetComponent implements OnInit, OnDestroy {
 
@@ -32,9 +29,9 @@ export class ForgotPasswordWidgetComponent implements OnInit, OnDestroy {
     this.submit$
       .withLatestFrom(this.mobileValidator$)
       .map(([_, mobileValidator]) => mobileValidator)
-      .filter<ValidT<MobileNumberT>>(mobileValidator => mobileValidator.isValid)
+      .filter(mobileValidator => mobileValidator.isValid)
       .subscribe(mobileValidator => {
-        const alias = mobileValidator.payload.mobileNumber;
+        const alias = (<ValidT<MobileNumberT>>mobileValidator).payload.mobileNumber;
         this.store.dispatch(new fromForgotPassword.ForgotPasswordRequestAction(alias));
         this.store.dispatch(new fromCredential.CredentialPutAliasAction(alias));
       });
