@@ -3,7 +3,6 @@ import { Observable } from 'rxjs/Observable';
 import { Subject } from 'rxjs/Subject';
 
 import { ValidationT, ValidT, MobileNumberT, PasswordT } from '../../services/form-validation/form-validation.service';
-import { Animations } from '../animations';
 
 import { Store } from '@ngrx/store';
 import { AppStateT } from '../../../../../ngrx';
@@ -14,7 +13,6 @@ import * as fromCredential from '../../../../../ngrx/auth/credential/credential.
   selector: 'msp-signup-widget',
   templateUrl: './signup-widget.component.html',
   styleUrls: ['./signup-widget.component.css'],
-  animations: [Animations.swipeInOut],
 })
 export class SignupWidgetComponent implements OnInit, OnDestroy {
 
@@ -23,6 +21,7 @@ export class SignupWidgetComponent implements OnInit, OnDestroy {
   submit$ = new Subject<true>();
   mobileValidator$ = new Subject<ValidationT<MobileNumberT>>();
   passwordValidator$ = new Subject<ValidationT<PasswordT>>();
+  //firstNameValidator$ = new Subject<ValidationT<FirstNameT>>();
 
   constructor(public store: Store<AppStateT>) { }
 
@@ -32,11 +31,11 @@ export class SignupWidgetComponent implements OnInit, OnDestroy {
     this.submit$
       .withLatestFrom(this.mobileValidator$, this.passwordValidator$)
       .map(([_, mobileValidator, passwordValidator]) => [mobileValidator, passwordValidator])
-      .filter<[ValidT<MobileNumberT>, ValidT<PasswordT>]>(([mobileValidator, passwordValidator]) =>
+      .filter(([mobileValidator, passwordValidator]) =>
         mobileValidator.isValid && passwordValidator.isValid)
       .subscribe(([mobileValidator, passwordValidator]) => {
-        const alias = mobileValidator.payload.mobileNumber;
-        const password = passwordValidator.payload.password;
+        const alias = (<ValidT<MobileNumberT>>mobileValidator).payload.mobileNumber;
+        const password = (<ValidT<PasswordT>>passwordValidator).payload.password;
         this.store.dispatch(new fromSignUp.SignUpRequestAction(alias, password));
         this.store.dispatch(new fromCredential.CredentialPutAliasAction(alias));
         this.store.dispatch(new fromCredential.CredentialPutPasswordAction(password));
