@@ -1,5 +1,5 @@
-import { Injectable } from '@angular/core';
-import { MobileContentService } from '../mobile-content/mobile-content.service';
+import {Injectable} from '@angular/core';
+import {MobileContentService} from '../mobile-content/mobile-content.service';
 
 
 export type ValidT<T> = {
@@ -26,10 +26,39 @@ export type MobileNumberT = {
   mobileNumber: string;
 };
 
+export type EmailT = {
+  email: string;
+};
+
+const emailRegex = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
+
 @Injectable()
 export class FormValidationService {
 
-  constructor(private mobileContentService: MobileContentService) { }
+  constructor(private mobileContentService: MobileContentService) {
+  }
+
+
+  emailValidator(email: string): ValidationT<EmailT> {
+    if (this.isEmpty(email)) {
+      return {
+        isValid: false,
+        errMsg: 'Required'
+      };
+    }
+
+    if (!email.match(emailRegex)) {
+      return {
+        isValid: false,
+        errMsg: 'Is an invalid email address.'
+      };
+    }
+
+    return {
+      isValid: true,
+      payload: {email: email},
+    };
+  }
 
   codeValidator(code: string): ValidationT<CodeT> {
     if (this.isEmpty(code)) {
@@ -37,7 +66,7 @@ export class FormValidationService {
         isValid: false,
         errMsg: 'Required'
       };
-    };
+    }
 
     if (!code.match(/[0-9]+/)) {
       return {
@@ -55,7 +84,7 @@ export class FormValidationService {
 
     return {
       isValid: true,
-      payload: { code: code }
+      payload: {code: code}
     };
   }
 
@@ -65,7 +94,8 @@ export class FormValidationService {
         isValid: false,
         errMsg: 'Required'
       };
-    };
+    }
+    ;
 
     if (!password.match(/^(?=.*[a-zA-Z])(?=.*[0-9])/)) {
       return {
@@ -83,7 +113,7 @@ export class FormValidationService {
 
     return {
       isValid: true,
-      payload: { password: password }
+      payload: {password: password}
     };
   }
 
@@ -93,7 +123,7 @@ export class FormValidationService {
         isValid: false,
         errMsg: 'Required'
       };
-    };
+    }
 
     const mobileContent = this.mobileContentService.getContent(countryNameCode);
 
@@ -114,11 +144,11 @@ export class FormValidationService {
 
     return {
       isValid: true,
-      payload: { mobileNumber: mobileNumber }
+      payload: {mobileNumber: mobileNumber}
     };
   }
 
   isEmpty(testString: string): boolean {
-    return testString.length === 0;
+    return !testString || testString.length === 0;
   }
 }
