@@ -1,5 +1,4 @@
-import { ActionReducer } from '@ngrx/store';
-import { environment } from '../../../environments/environment';
+import {ActionReducer, ActionReducerMap} from '@ngrx/store';
 
 import 'rxjs/add/operator/catch';
 import 'rxjs/add/operator/map';
@@ -9,23 +8,6 @@ import 'rxjs/add/operator/switchMap';
 import 'rxjs/add/operator/switchMapTo';
 
 /**
- * The compose function is one of our most handy tools. In basic terms, you give
- * it any number of functions and it returns a function. This new function
- * takes a value and chains it through every composed function, returning
- * the output.
- *
- * More: https://drboolean.gitbooks.io/mostly-adequate-guide/content/ch5.html
- */
-import { compose } from '@ngrx/core/compose';
-
-/**
- * storeFreeze prevents state from being mutated. When mutation occurs, an
- * exception will be thrown. This is useful during development mode to
- * ensure that none of the reducers accidentally mutates the state.
- */
-// import { storeFreeze } from 'ngrx-store-freeze';
-
-/**
  * combineReducers is another useful metareducer that takes a map of reducer
  * functions and creates a new reducer that stores the gathers the values
  * of each reducer and stores them using the reducer's key. Think of it
@@ -33,7 +15,7 @@ import { compose } from '@ngrx/core/compose';
  *
  * More: https://egghead.io/lessons/javascript-redux-implementing-combinereducers-from-scratch
  */
-import { combineReducers } from '@ngrx/store';
+import {combineReducers} from '@ngrx/store';
 
 /**
  * Every reducer module's default export is the reducer function itself. In
@@ -41,32 +23,28 @@ import { combineReducers } from '@ngrx/store';
  * the state of the reducer plus any selector functions. The `* as`
  * notation packages up all of the exports into a single object.
  */
-// import * as fromAuthLogin from './auth/login/login.reducer';
-
-import { credentialReducer, CredentialStateT } from './credential/credential.reducer';
-import { currentReducer, CurrentStateT } from './current/current.reducer';
-import { resendCodeReducer, ResendCodeStateT } from './resend-code/resend-code.reducer';
-import { signUpReducer, SignUpStateT } from './sign-up/sign-up.reducer';
-import { confirmRegistrationReducer, ConfirmRegistrationStateT } from './confirm-registration/confirm-registration.reducer';
-import { loginReducer, LoginStateT } from './login/login.reducer';
-import { confirmNewPasswordReducer, ConfirmNewPasswordStateT } from './confirm-new-password/confirm-new-password.reducer';
-import { forgotPasswordReducer, ForgotPasswordStateT } from './forgot-password/forgot-password.reducer';
-
-// import { reducer as loginReducer, LoginStateT } from './login/login.reducer';
+import * as fromCredential from './credential/credential.reducer';
+import * as fromCurrent from './current/current.reducer';
+import * as fromResendCode from './resend-code/resend-code.reducer';
+import * as fromSignUp from './sign-up/sign-up.reducer';
+import * as fromConfirmRegistration from './confirm-registration/confirm-registration.reducer';
+import * as fromLogin from './login/login.reducer';
+import * as fromConfirmNewPassword from './confirm-new-password/confirm-new-password.reducer';
+import * as fromForgotPassword from './forgot-password/forgot-password.reducer';
 
 /**
  * As mentioned, we treat each reducer like a table in a database. This means
  * our top level state interface is just a map of keys to inner state types.
  */
 export type AuthStateT = {
-  credential: CredentialStateT,
-  current: CurrentStateT,
-  resendCode: ResendCodeStateT,
-  signUp: SignUpStateT,
-  confirmRegistration: ConfirmRegistrationStateT,
-  login: LoginStateT,
-  forgotPassword: ForgotPasswordStateT,
-  confirmNewPassword: ConfirmNewPasswordStateT
+  credential: fromCredential.CredentialStateT,
+  current: fromCurrent.CurrentStateT,
+  resendCode: fromResendCode.ResendCodeStateT,
+  signUp: fromSignUp.SignUpStateT,
+  confirmRegistration: fromConfirmRegistration.ConfirmRegistrationStateT,
+  login: fromLogin.LoginStateT,
+  forgotPassword: fromForgotPassword.ForgotPasswordStateT,
+  confirmNewPassword: fromConfirmNewPassword.ConfirmNewPasswordStateT
 };
 
 /**
@@ -76,25 +54,30 @@ export type AuthStateT = {
  * wrapping that in storeLogger. Remember that compose applies
  * the result from right to left.
  */
-const authReducers = {
-  credential: credentialReducer,
-  current: currentReducer,
-  resendCode: resendCodeReducer,
-  signUp: signUpReducer,
-  confirmRegistration: confirmRegistrationReducer,
-  login: loginReducer,
-  forgotPassword: forgotPasswordReducer,
-  confirmNewPassword: confirmNewPasswordReducer,
+export const authReducers: ActionReducerMap<AuthStateT> = {
+  credential: fromCredential.credentialReducer,
+  current: fromCurrent.currentReducer,
+  resendCode: fromResendCode.resendCodeReducer,
+  signUp: fromSignUp.signUpReducer,
+  confirmRegistration: fromConfirmRegistration.confirmRegistrationReducer,
+  login: fromLogin.loginReducer,
+  forgotPassword: fromForgotPassword.forgotPasswordReducer,
+  confirmNewPassword: fromConfirmNewPassword.confirmNewPasswordReducer,
 };
 
-// const authDevReducer: ActionReducer<AuthStateT> = compose(storeFreeze, combineReducers)(authReducers);
 const authProReducer: ActionReducer<AuthStateT> = combineReducers(authReducers);
 
-export function authReducer(state: AuthStateT, action: any) {
+export const initialState: AuthStateT = {
+  credential: fromCredential.initialState,
+  current: fromCurrent.initialState,
+  resendCode: fromResendCode.initialState,
+  signUp: fromSignUp.initialState,
+  confirmRegistration: fromConfirmRegistration.initialState,
+  login: fromLogin.initialState,
+  forgotPassword: fromForgotPassword.initialState,
+  confirmNewPassword: fromConfirmNewPassword.initialState,
+};
+
+export function authReducer(state: AuthStateT = initialState, action: any) {
   return authProReducer(state, action);
-  // if (environment.production) {
-  //   return authProReducer(state, action);
-  // } else {
-  //   return authDevReducer(state, action);
-  // }
 }
