@@ -1,22 +1,4 @@
-import { ActionReducer } from '@ngrx/store';
-import { environment } from '../../environments/environment';
-
-/**
- * The compose function is one of our most handy tools. In basic terms, you give
- * it any number of functions and it returns a function. This new function
- * takes a value and chains it through every composed function, returning
- * the output.
- *
- * More: https://drboolean.gitbooks.io/mostly-adequate-guide/content/ch5.html
- */
-import { compose } from '@ngrx/core/compose';
-
-/**
- * storeFreeze prevents state from being mutated. When mutation occurs, an
- * exception will be thrown. This is useful during development mode to
- * ensure that none of the reducers accidentally mutates the state.
- */
-// import { storeFreeze } from 'ngrx-store-freeze';
+import {ActionReducer, ActionReducerMap} from '@ngrx/store';
 
 /**
  * combineReducers is another useful metareducer that takes a map of reducer
@@ -34,16 +16,14 @@ import { combineReducers } from '@ngrx/store';
  * the state of the reducer plus any selector functions. The `* as`
  * notation packages up all of the exports into a single object.
  */
-import { AuthStateT, authReducer } from './auth';
-// import { InitStateT, initReducer } from './init/init.reducer';
+import * as fromAuth from './auth';
 
 /**
  * As mentioned, we treat each reducer like a table in a database. This means
  * our top level state interface is just a map of keys to inner state types.
  */
 export interface AppStateT {
-  // init: InitStateT;
-  auth: AuthStateT;
+  auth: fromAuth.AuthStateT;
 }
 
 /**
@@ -53,19 +33,16 @@ export interface AppStateT {
  * wrapping that in storeLogger. Remember that compose applies
  * the result from right to left.
  */
-const appReducers = {
-  // init: initReducer,
-  auth: authReducer,
+export const appReducers: ActionReducerMap<AppStateT> = {
+  auth: fromAuth.authReducer,
 };
 
-// const developmentReducer: ActionReducer<AppStateT> = compose(storeFreeze, combineReducers)(appReducers);
 const productionReducer: ActionReducer<AppStateT> = combineReducers(appReducers);
 
-export function appReducer(state: AppStateT, action: any) {
+export const initialState: AppStateT = {
+  auth: fromAuth.initialState
+};
+
+export function appReducer(state: AppStateT = initialState, action: any) {
   return productionReducer(state, action);
-  // if (environment.production) {
-  //   return productionReducer(state, action);
-  // } else {
-  //   return developmentReducer(state, action);
-  // }
 }
